refactor(catalog): split the m3 sticker sheet into per-cluster files
`CatalogPreviews.kt` was 484 lines and ~40 previews in one file, which is
one file to open for any change and — since the playground handoff landed
— the whole file a visitor gets when they click "open in playground" on a
single badge.

Split along the section comments that were already there, which line up
with the `group =` values on the annotations:

    CatalogButtons.kt        57   Buttons
    CatalogSelection.kt      62   Selection controls
    CatalogContainment.kt    55   Cards + FAB
    CatalogCommunication.kt  22   Progress + badge
    CatalogTextFields.kt     17   Text fields
    CatalogText.kt           38   Text options
    CatalogStates.kt         50   Interaction states + content axis
    CatalogI18n.kt           83   i18n / a11y axes
    CatalogTemplates.kt      94   Scaffold templates
    CatalogStickers.kt       53   The sticker-sheet doc + the Sticker helper

Pure move: every preview function, annotation and comment is carried over
verbatim. The only edits are mechanical — `Sticker` becomes `internal` so
the cluster files can call it, each file gets the subset of imports its
own body uses, and `CatalogTemplates.kt` carries the `@file:OptIn` the
original header held for it.

Verified by rendering: 82 PNGs before and after, same names.
